BJS2-66141: правки
