+++ src/main/java/school/faang/user_service/mapper/recommendation/SkillRequestMapper.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.mapper;
+package school.faang.user_service.mapper.recommendation;
 
 import org.mapstruct.InheritInverseConfiguration;
 import org.mapstruct.Mapper;
