+++ src/main/java/school/faang/user_service/service/recommendation/RecommendationRequestService.java
@@ -11,8 +11,8 @@ import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
 import school.faang.user_service.exception.NotFoundRequestException;
-import school.faang.user_service.filter.recommendation.RecommendationRequestFilterProcessor;
-import school.faang.user_service.mapper.RecommendationRequestMapper;
+import school.faang.user_service.filter.recommendationrequest.RecommendationRequestFilterProcessor;
+import school.faang.user_service.mapper.recommendation.RecommendationRequestMapper;
 import school.faang.user_service.repository.SkillRepository;
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.recommendation.RecommendationRequestRepository;
