+++ src/test/java/school/faang/user_service/filter/recommendation/RecommendationRequestRequesterFilterTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.recommendation.RequestFilterDto;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.recommendationrequest.RecommendationRequestRequesterFilter;
 
 import java.util.List;
