+++ src/test/java/school/faang/user_service/filter/recommendation/RecommendationRequestStatusFilterTest.java
@@ -4,6 +4,7 @@ import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.recommendation.RequestFilterDto;
 import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.recommendationrequest.RecommendationRequestStatusFilter;
 
 import java.util.Arrays;
 import java.util.List;
