+++ src/test/java/school/faang/user_service/filter/recommendation/RecommendationRequestUpdatedAtFilterTest.java
@@ -3,6 +3,7 @@ package school.faang.user_service.filter.recommendation;
 import org.junit.jupiter.api.Test;
 import school.faang.user_service.dto.recommendation.RequestFilterDto;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
+import school.faang.user_service.filter.recommendationrequest.RecommendationRequestUpdatedAtFilter;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
