+++ src/test/java/school/faang/user_service/service/recommendation/RecommendationRequestServiceTest.java
@@ -15,7 +15,7 @@ import school.faang.user_service.entity.RequestStatus;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.recommendation.RecommendationRequest;
 import school.faang.user_service.exception.NotFoundRequestException;
-import school.faang.user_service.mapper.RecommendationRequestMapper;
+import school.faang.user_service.mapper.recommendation.RecommendationRequestMapper;
 import school.faang.user_service.repository.SkillRepository;
 import school.faang.user_service.repository.UserRepository;
 import school.faang.user_service.repository.recommendation.RecommendationRequestRepository;
